Migrate ProductType + ProductOption to split-class layout
ProductTypeResource     163 -> 65 lines  (+ProductTypeForm 67, +ProductTypeTable 60)
ProductOptionResource   153 -> 65 lines  (+ProductOptionForm 80, +ProductOptionTable 60)

Both mechanical; no new patterns surfaced.
